ugc_api/pyproject.toml
[project]
name = "ugc-api"
version = "0.1.0"
description = "UGC API with MongoDB"
requires-python = ">=3.11"
dependencies = [
    "aiokafka==0.9.0",
    "beanie>=1.29.0",
    "fastapi[standard]>=0.115.7",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "orjson>=3.10.15",
    "pydantic-settings>=2.7.1",
    "pyjwt[crypto]>=2.10.1",
    "sentry-sdk>=2.20.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "mongomock-motor>=0.0.25",
    "httpx>=0.27.0",
    "asgi-lifespan>=2.1.0",
    "dependency-injector>=4.41.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/tests"]
python_files = ["test_*.py"]
addopts = "-v -s --tb=short"
